test: disable calibration sleep timers if there is no hardware
The sleep timers are needed if a physical fan is connected. They allow the
fan to react to the speed changes and settle in on the new RPM. If there is
no fan connected and we're testing against an emulation then we don't need
to wait for anything to settle.